% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions_package.R
\docType{data}
\name{BCI.env.coarse}
\alias{BCI.env.coarse}
\title{Barro Colorado Island 50 Hectare Quadrat Environmental Dataset at Coarse Scale}
\format{A data frame with 50 rows (quadrats) and 15 columns (environmental variables).}
\usage{
data(BCI.env.coarse)
}
\description{
Measurements of soil composition and elevation in 50, 100m x 100m contiguous quadrats.
}
\details{
The data set contains the following variables:

\itemize{
\item{x}: x-position in meters
\item{y}: y-position in meters
\item{Al}: Aluminium concentration
\item{B}: Boron concentration
\item{Ca}: Calcium concentration
\item{Cu}: Copper concentration
\item{Fe}: Iron concentration
\item{K}: Potassium concentration
\item{Mg}: Magnesium concentration
\item{Mn}: Manganese concentration
\item{P}: Phosphorous concentration
\item{Zn}: Zinc concentration
\item{N}: Nitrogen concentration
\item{pH}: Interpolated pH content in the surface soil
\item{elevation}: mean elevation
}

Location: Panama -- 9 degrees 9' 4.5" N, 79 degrees 51' 19.08" W

Data owner: Center for Tropical Forest Science of the Smithsonian Tropical Research Institute
}
\references{
Hubbell, S.P., Condit, R., and Foster, R.B. (2005). Barro Colorado Forest Census Plot Data. \url{http://ctfs.arnarb.harvard.edu/webatlas/datasets/bci/}

Condit, R. (1998). \emph{Tropical Forest Census Plots}. Springer-Verlag and R. G. Landes Company, Berlin, Germany, and Georgetown, Texas.

Hubbell, S.P., R.B. Foster, S.T. O'Brien, K.E. Harms, R. Condit, B. Wechsler, S.J. Wright, and S. Loo de Lao. (1999). Light gap disturbances, recruitment limitation, and tree diversity in a neotropical forest. \emph{Science} 283, 554-557.

\url{http://ctfs.arnarb.harvard.edu/webatlas/datasets/bci/soilmaps/BCIsoil.html}
}
\keyword{data}

