% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetadiv.R
\docType{data}
\name{bird.env.fine}
\alias{bird.env.fine}
\title{South-East Australia Environmental Dataset at Fine Scale}
\format{
A data frame with 604 rows (sites) and 9 columns (xy coordinates and environmental variables).
}
\usage{
data(bird.env.fine)
}
\description{
Projected coordinates and environmental variables in 604, 25 x 25 km contiguous sites.
}
\details{
The data set contains the following variables:

\itemize{
\item{x}: x-position in meters in UTM 53 South projection
\item{y}: y-position in meters in UTM 53 South projection
\item{Natural}: Proportion of area of conservation and natural environments
\item{Irrigated}: Proportion of area of production from irrigated agriculture and plantations
\item{Water}: Proportion of area of water features
\item{Elevation}: Elevation
\item{ApP}: Area per person
\item{Temp}: Temperature
\item{Precip}: Precipitation
}

Location: Australia -- 50° 33' 5.03" S, 135° 21' 10.40" E

Data owners: ABARES, Australian Bureau of Statistics,GEBCO, WorldClim
}
\references{
http://data.daff.gov.au/anrdl/metadata_files/pa_luav4g9abl07811a00.xml

http://www.gebco.net/

http://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/1270.0.55.0072011?

http://www.worldclim.org/

Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G. & Jarvis, A. (2005) Very high resolution interpolated climate surfaces for global land areas. International journal of climatology, 25, 1965-1978.
}
\keyword{data}
