% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetadiv.R
\docType{data}
\name{bird.spec.coarse}
\alias{bird.spec.coarse}
\title{Australia Bird Atlas Species Occurrence Dataset at Coarse Scale over South-East Australia}
\format{A data frame with 123 rows (sites) and 193 columns (xy coordinates and species).}
\usage{
data(bird.spec.coarse)
}
\description{
Inventory of bird species occurrence in 123, 100 x 100 km sites.
}
\details{
\itemize{
\item{x}: x-position in meters in UTM 53 South projection
\item{y}: y-position in meters in UTM 53 South projection
\item{columns 3-193}: bird species occurrence
}

The original bird occurrence data were arranged into a continuous grid covering South-East Australia. Only cells whose richness was within 10 percents of real estimated richness are included here, so that the data corresponds to presence-absence data.

Location: Australia -- 51° 27' 2.27" S, 135° 21' 35.19" E

Data owner: BirdLife Australia
}
\references{
Barrett, G., Silcocks, A., Barry, S., Cunningham, R. & Poulter, R. (2003) The new atlas of Australian birds. Royal Australasian Ornithologists Union, Melbourne, 1-824.
}
\keyword{data}

