% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetadiv.R
\docType{data}
\name{bird.spec.fine}
\alias{bird.spec.fine}
\title{Australia Bird Atlas Species Occurrence Dataset at Fine Scale over South-East Australia}
\format{A data frame with 604 rows (sites) and 193 columns (xy coordinates and species).}
\usage{
data(bird.spec.fine)
}
\description{
Inventory of bird species occurrence in 604, 25 x 25 km sites.
}
\details{
\itemize{
\item{x}: x-position in meters in UTM 53 South projection
\item{y}: y-position in meters in UTM 53 South projection
\item{columns 3-192}: bird species occurrence
}

Location: Australia -- 50° 33' 5.03" S, 135° 21' 10.40" E

Data owner: BirdLife Australia

The original bird occurrence data were arranged into a continuous grid covering South-East Australia. Only cells whose richness was within 10 percents of real estimated richness are included here, so that the data corresponds to presence-absence data.
}
\references{
Barrett, G., Silcocks, A., Barry, S., Cunningham, R. & Poulter, R. (2003) The new atlas of Australian birds. Royal Australasian Ornithologists Union, Melbourne, 1-824.
}
\keyword{data}

