% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions_package.R
\docType{data}
\name{Marion.env}
\alias{Marion.env}
\title{Marion Island Environmental Dataset}
\format{A data frame with 12 rows (plots) and 4 columns (variables).}
\usage{
data(Marion.env)
}
\description{
Geographic coordinates, altitude and island side (East, West) at 12 plots (4 transects and 3 altitudes) on Marion Island.
}
\details{
The data set contains the following variables:

\itemize{
\item{x}: x-position in meters in UTM37G projection
\item{y}: y-position in meters in UTM37G projection
\item{Altitude}: mean elevation
\item{Side}: cardinal (East or West) side of the island
}

Location: Marion Island -- 46 degrees 53' 34.2" S, 37 degrees 45' 02.3" E

Data owner: Melodie A. McGeoch
}
\references{
Nyakatya, M.J. & McGeoch, M.A. (2008). Temperature variation across Marion Island associated with a keystone plant species (\emph{Azorella selago} Hook. (Apiaceae)). Polar Biology, 31, 139-151.

McGeoch, M.A., Le Roux, P.C., Hugo, E.A. & Nyakatya, M.J. (2008). Spatial variation in the terrestrial biotic system. The Prince Edward Islands: Land-Sea Interactions in a Changing World (ed. by S.L. Chown and P.W. Froneman), pp. 245-276. African SunMedia, Stellenbosch.
}
\keyword{data}

