% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetadiv.R
\docType{data}
\name{Marion.species}
\alias{Marion.species}
\title{Marion Island Species Presence-Absence Dataset}
\format{A data frame with 12 rows (plots) and 33 columns (species).}
\usage{
data(Marion.species)
}
\description{
Inventory of springtails and mite species presence-absence in 12 plots (4 transects and 3 altitudes) on Marion Island.
}
\details{
The data set contains the following variables:

\itemize{
\item{x}: x-position in meters in UTM 37 South projection
\item{y}: y-position in meters in UTM 37 South projection
\item{columns 3-24}: mite species presence absence
\item{columns 25-33}: springtail species presence absence
}

Location: Marion Island -- 46° 53' 34.2" S, 37 degrees 45' 02.3" E

Data owner: Melodie A. McGeoch
}
\references{
Nyakatya, M.J. & McGeoch, M.A. (2008). Temperature variation across Marion Island associated with a keystone plant species (\emph{Azorella selago} Hook. (Apiaceae)). Polar Biology, 31, 139-151.

McGeoch, M.A., Le Roux, P.C., Hugo, E.A. & Nyakatya, M.J. (2008). Spatial variation in the terrestrial biotic system. The Prince Edward Islands: Land-Sea Interactions in a Changing World (ed. by S.L. Chown and P.W. Froneman), pp. 245-276. African SunMedia, Stellenbosch.
}
\keyword{data}
